+++ packages/guide/src/app/shared/shared.module.ts
@@ -24,6 +24,10 @@ import { ResponsiveTableComponent } from './components/responsive-table/responsi
 import { MatTableModule } from '@angular/material/table';
 import { SpacesToPascalCasePipe } from './pipes/spaces-to-pascal-case.pipe';
 import { IsCropPipe } from './pipes/is-crop.pipe';
+import { RefinementsListComponent } from './components/refinements-list/refinements-list.component';
+import { CastToQualityPipe } from './pipes/cast-to-quality.pipe';
+import { ProcessingTimeComponent } from './components/processing-time/processing-time.component';
+import { ProcessingTimePerQualityComponent } from './components/processing-time-per-quality/processing-time-per-quality.component';
 
 @NgModule({
     declarations: [
@@ -44,6 +48,10 @@ import { IsCropPipe } from './pipes/is-crop.pipe';
         ResponsiveTableComponent,
         SpacesToPascalCasePipe,
         IsCropPipe,
+        RefinementsListComponent,
+        CastToQualityPipe,
+        ProcessingTimeComponent,
+        ProcessingTimePerQualityComponent,
     ],
     imports: [
         CommonModule,
@@ -72,6 +80,11 @@ import { IsCropPipe } from './pipes/is-crop.pipe';
         ResponsiveTableComponent,
         SpacesToPascalCasePipe,
         IsCropPipe,
+        RefinementsListComponent,
+        RarityIconComponent,
+        CastToQualityPipe,
+        ProcessingTimeComponent,
+        ProcessingTimePerQualityComponent,
     ],
 })
 export class SharedModule {